tp2/requirements.txt
aiohttp==3.8.1
Pillow==9.2.0
argparse==1.4.0